split pre-release to own controller
